+++ bundles/statistics/bundle/statsgrid/GridModeView.js
@@ -69,8 +69,10 @@ function() {
 
             // Notify other components of the mode change.
             var eventBuilder = this.instance.getSandbox().getEventBuilder('StatsGrid.ModeChangedEvent');
-            var event = eventBuilder(this.isVisible);
-            this.instance.getSandbox().notifyAll(event);
+            if (eventBuilder) {
+                var event = eventBuilder(this.isVisible);
+                this.instance.getSandbox().notifyAll(event);
+            }
         }
     },