+++ ReleaseNotes.md
@@ -10,6 +10,9 @@ mapOptions configuration can now be used to set units for OpenLayers.Map (defaul
 
 Scale limitations now use map resolutions internally to minimize risk of scale/resolution transformation errors.
 
+### ManageStatsPlugin
+There is now a possibility to uncheck some of the municipalities. This affects to the statistical variables. This feature can be swithced on from header row drop down list
+
 ## 1.11
 
 ### core/sandbox