+++ ReleaseNotes.md
@@ -1,5 +1,17 @@
 # Release Notes
 
+## 1.12
+
+### statsgrid bundle
+
+Highlight/select controls are now disabled when not in the stats mode.
+
+Clicking on an area on the map highlights the corresponding row in the grid and scrolls to display it as the topmost row.
+
+### mapstats bundle
+
+Hovering over an area on the map sends a request to get tooltip info which is then shown over the area.
+
 ## 1.11
 
 ### core/sandbox